requirements.txt
# Otonom AI Şirketi - Gerekli Paketler
# Python 3.8+ gereklidir

# Core AI ve LangChain
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.20
anthropic>=0.8.0
google-generativeai>=0.3.0

# Veri ve Yapılandırma
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1

# Web Framework - FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Web UI - Streamlit
streamlit>=1.31.0
streamlit-autorefresh>=1.0.0

# Data Analysis ve Visualization
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
plotly>=5.18.0
altair>=5.2.0

# Database (Opsiyonel)
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
redis>=5.0.1

# Task Queue (Opsiyonel)
celery>=5.3.6

# Terminal UI
colorama>=0.4.6
rich>=13.7.0

# Utilities
schedule>=1.2.0
asyncio>=3.4.3
requests>=2.31.0
httpx>=0.26.0
jinja2>=3.1.3
python-dateutil>=2.8.2

# Testing (Opsiyonel)
pytest>=7.4.0
pytest-asyncio>=0.23.0
